feat: remove unused CameraOverlayView, SnapTips, and ExamplePhotos components
